+++ NEWS.md
@@ -2,4 +2,5 @@
 
 * Initial package development
 * Add `get_pvol` function to download polar volumes for 6 countries
-* Add `get_vpts` function to download volume scans from the [Aloft data repository](https://aloftdata.eu/)
+* Add `get_vpts` function to download vertical profile time series from the [Aloft data repository](https://aloftdata.eu/)
+* Add `aloft_data_coverage` function to fetch an overview table of the files available on the [Aloft data repository](https://aloftdata.eu/)